// src/utils/enums/Hindrance.tsx
export enum Hindrance {
  AilinMinor = "AilinMinor",
  AilinMajor = "AilinMajor",
  AllThumbs = "AllThumbs",
  Arrogant = "Arrogant",
  BadDreams = "BadDreams",
  BadEyes = "BadEyes",
  BadLuck = "BadLuck",
  BigMouth = "BigMouth",
  Blind = "Blind",
  Bloodthirsty = "Bloodthirsty",
  Cautious = "Cautious",
  Clueless = "Clueless",
  CodeOfHonor = "CodeOfHonor",
  Curious = "Curious",
  Cursed = "Cursed",
  DeathWish = "DeathWish",
  DelusionalMinor = "DelusionalMinor",
  DelusionalMajor = "DelusionalMajor",
  Elderly = "Elderly",
  EnemyMinor = "EnemyMinor",
  EnemyMajor = "EnemyMajor",
  FocusMajor = "FocusMajor",
  FocusMinor = "FocusMinor",
  GreedyMinor = "GreedyMinor",
  GreedyMajor = "GreedyMajor",
  GrimServant = "GrimServant",
  HabitMinor = "HabitMinor",
  HabitMajor = "HabitMajor",
  HardOfHearingMinor = "HardOfHearingMinor",
  HardOfHearingMajor = "HardOfHearingMajor",
  HeavySleeper = "HeavySleeper",
  Heroic = "Heroic",
  Illiterate = "Illiterate",
  Lame = "Lame",
  Loyal = "Loyal",
  LyingEyes = "LyingEyes",
  Mean = "Mean",
  Obese = "Obese",
  OldWaysOathMinor = "OldWaysOathMinor",
  OldWaysOathMajor = "OldWaysOathMajor",
  OneArm = "OneArm",
  OneEye = "OneEye",
  OneLeg = "OneLeg",
  Outsider = "Outsider",
  Overconfident = "Overconfident",
  PacifistMinor = "PacifistMinor",
  PacifistMajor = "PacifistMajor",
  PhobiaMinor = "PhobiaMinor",
  PhobiaMajor = "PhobiaMajor",
  Poverty = "Poverty",
  Quirk = "Quirk",
  Slowpoke = "Slowpoke",
  Small = "Small",
  Stubborn = "Stubborn",
  Tenderfoot = "Tenderfoot",
  ThinSkinned = "ThinSkinned",
  TroubleMagnet = "TroubleMagnet",
  Ugly = "Ugly",
  VengefulMinor = "VengefulMinor",
  VengefulMajor = "VengefulMajor",
  Vow = "Vow",
  WantedMinor = "WantedMinor",
  WantedMajor = "WantedMajor",
  Yellow = "Yellow",
  Young = "Young",
}